test(senpi-task): split category boundary tests
