+++ .omo/evidence/senpi-task/task-4-category/repair4-test-split/01-category-tests.log
@@ -1,0 +1,26 @@
+bun test v1.3.14 (0d9b296a)
+
+packages/senpi-task/src/category/resolve-category.test.ts:
+(pass) resolveCategory > #given a builtin category and omo overlay #when resolved #then user config wins and prompt text is appended [0.94ms]
+(pass) resolveCategory > #given a disabled omo category overlay #when resolved #then a disabled result explains the reason [0.08ms]
+(pass) resolveCategory > #given primary model is unavailable and omo fallback exists #when resolved #then delegate-core fallback reaches the registry model [0.22ms]
+(pass) resolveCategory > #given quick primary is unavailable and hardcoded fallback is available #when resolved #then delegate-core fallback chain reaches Anthropic Haiku [0.10ms]
+(pass) resolveCategory > #given ultrabrain primary is unavailable and hardcoded Google fallback is available #when resolved #then delegate-core fallback chain preserves the high variant [0.09ms]
+(pass) resolveCategory > #given no category or fallback model resolves and a system default is available #when resolved #then delegate-core reaches the system default [0.14ms]
+(pass) resolveCategory > #given selected model is absent from registry #when resolved #then unavailable result names attempted and available models [0.09ms]
+(pass) resolveCategory > #given category params in omo overlay #when resolved #then child spec carries generation params and prompt append [0.09ms]
+(pass) resolveCategory > #given a custom category description #when resolved #then the resolved result preserves it [0.06ms]
+(pass) builtin category defaults > #given ported builtin defaults #when snapshotted #then all eight category defaults stay pinned [0.26ms]
+
+packages/senpi-task/src/category/resolve-category-boundary.test.ts:
+(pass) resolveCategory boundary parsing > #given a registry model with legal headers #when resolved #then the header-bearing model is accepted [0.44ms]
+(pass) resolveCategory boundary parsing > #given a malformed registry entry #when resolved #then category resolution returns sanitized model_unavailable instead of throwing [0.08ms]
+(pass) resolveCategory boundary parsing > #given malformed truthy find results #when resolved #then category resolution returns sanitized model_unavailable [0.13ms]
+(pass) resolveCategory boundary parsing > #given inherited model identity fields #when resolved #then category resolution rejects them without leaking prototype data [0.07ms]
+(pass) resolveCategory boundary parsing > #given non-array registry availability #when resolved #then category resolution returns sanitized model_unavailable instead of throwing [0.07ms]
+(pass) resolveCategory boundary parsing > #given prototype-shaped category names #when resolved #then they return not_found instead of inherited object values [0.07ms]
+
+ 16 pass
+ 0 fail
+ 1 snapshots, 84 expect() calls
+Ran 16 tests across 2 files. [61.00ms]
