+++ .omo/evidence/senpi-task/task-4-category/repair4-test-split/02-senpi-task-bail.exit
@@ -1,0 +1,1 @@
+0
