+++ .omo/evidence/senpi-task/task-4-category/repair4-test-split/02-senpi-task-bail.log
@@ -1,0 +1,69 @@
+bun test v1.3.14 (0d9b296a)
+
+packages/senpi-task/src/senpi-api-tripwire.test.ts:
+(pass) pinned Senpi API surface > #given senpi root exports #when type checked #then task adapter can pass session construction seams [0.72ms]
+(pass) pinned Senpi API surface > #given agent dir marker extension #when session boots with minimal loader #then marker factory is not invoked [499.43ms]
+(pass) pinned Senpi API surface > #given minimal resource loader source #when audited #then fake marker factory option is absent [0.11ms]
+(pass) pinned Senpi API surface > #given pinned artifact #when package metadata and rpc entry are checked #then expected public contract exists [0.24ms]
+
+packages/senpi-task/src/category/resolve-category.test.ts:
+(pass) resolveCategory > #given a builtin category and omo overlay #when resolved #then user config wins and prompt text is appended [0.66ms]
+(pass) resolveCategory > #given a disabled omo category overlay #when resolved #then a disabled result explains the reason [0.09ms]
+(pass) resolveCategory > #given primary model is unavailable and omo fallback exists #when resolved #then delegate-core fallback reaches the registry model [0.21ms]
+(pass) resolveCategory > #given quick primary is unavailable and hardcoded fallback is available #when resolved #then delegate-core fallback chain reaches Anthropic Haiku [0.09ms]
+(pass) resolveCategory > #given ultrabrain primary is unavailable and hardcoded Google fallback is available #when resolved #then delegate-core fallback chain preserves the high variant [0.08ms]
+(pass) resolveCategory > #given no category or fallback model resolves and a system default is available #when resolved #then delegate-core reaches the system default [0.09ms]
+(pass) resolveCategory > #given selected model is absent from registry #when resolved #then unavailable result names attempted and available models [0.07ms]
+(pass) resolveCategory > #given category params in omo overlay #when resolved #then child spec carries generation params and prompt append [0.10ms]
+(pass) resolveCategory > #given a custom category description #when resolved #then the resolved result preserves it [0.06ms]
+(pass) builtin category defaults > #given ported builtin defaults #when snapshotted #then all eight category defaults stay pinned [0.27ms]
+
+packages/senpi-task/src/category/resolve-category-boundary.test.ts:
+(pass) resolveCategory boundary parsing > #given a registry model with legal headers #when resolved #then the header-bearing model is accepted [0.84ms]
+(pass) resolveCategory boundary parsing > #given a malformed registry entry #when resolved #then category resolution returns sanitized model_unavailable instead of throwing [0.07ms]
+(pass) resolveCategory boundary parsing > #given malformed truthy find results #when resolved #then category resolution returns sanitized model_unavailable [0.15ms]
+(pass) resolveCategory boundary parsing > #given inherited model identity fields #when resolved #then category resolution rejects them without leaking prototype data [0.08ms]
+(pass) resolveCategory boundary parsing > #given non-array registry availability #when resolved #then category resolution returns sanitized model_unavailable instead of throwing [0.08ms]
+(pass) resolveCategory boundary parsing > #given prototype-shaped category names #when resolved #then they return not_found instead of inherited object values [0.07ms]
+
+packages/senpi-task/src/state/store.test.ts:
+(pass) resolveStateDir > #given no task state override #when resolved #then project omo senpi-task directory is used [0.13ms]
+(pass) resolveStateDir > #given task state override #when resolved #then override directory wins [0.05ms]
+(pass) TaskRecordStore > #given completed task record #when saved and reloaded #then durable task facts are identical [1.09ms]
+(pass) TaskRecordStore > #given sensitive event payload #when event is appended #then jsonl payload is redacted [0.82ms]
+(pass) TaskRecordStore > #given corrupt task json #when records are listed #then typed diagnostic is reported and corrupt record is skipped [0.71ms]
+(pass) TaskRecordStore > #given duplicate generated task id #when a different record is saved #then existing task file is not overwritten [0.61ms]
+(pass) TaskRecordStore > #given completed record #when illegal running transition is persisted #then transition is rejected and completion remains [0.66ms]
+
+packages/senpi-task/src/state/transitions-table.test.ts:
+(pass) transitionTaskRecord exhaustive lifecycle table > #given every status and normal transition #when applied #then allowed and rejected outcomes match the lifecycle table [0.38ms]
+
+packages/senpi-task/src/state/model.test.ts:
+(pass) messageability > #given every status and residency pair #when classified #then the table is exhaustive [0.17ms]
+(pass) transitionTaskRecord > #given a cancelled task #when late failure arrives #then cancelled remains terminal and failure is logged [0.09ms]
+
+packages/senpi-task/src/state/id.test.ts:
+(pass) createTaskId > #given a deterministic clock #when ids are created #then ids are canonical and sortable [0.73ms]
+(pass) createTaskId > #given a same-millisecond burst #when more than one byte of ids are created #then every id is unique [0.23ms]
+(pass) createTaskId > #given an isolated deterministic clock #when ids are created #then the clock seam is repeatable [0.09ms]
+(pass) createTaskId > #given wrap pressure near the id ceiling #when ids are created #then ids never wrap or repeat [0.22ms]
+
+packages/senpi-task/src/state/transitions.test.ts:
+(pass) transitionTaskRecord lifecycle graph > #given pending task #when terminal transitions arrive before start #then they are rejected [0.19ms]
+(pass) transitionTaskRecord lifecycle graph > #given running task #when terminal transitions arrive #then the lifecycle terminal is applied [0.09ms]
+(pass) transitionTaskRecord lifecycle graph > #given pending or running task #when normal lose transition arrives #then lost is rejected [0.07ms]
+(pass) transitionTaskRecord lifecycle graph > #given pending or running task #when reconciliation marks lost #then lost is applied explicitly [0.07ms]
+(pass) transitionTaskRecord lifecycle graph > #given terminal task #when residency-only transition arrives #then status remains and residency changes [0.07ms]
+(pass) transitionTaskRecord lifecycle graph > #given every terminal status #when every residency transition arrives #then only residency changes [0.14ms]
+(pass) transitionTaskRecord lifecycle graph > #given interrupted task #when late completion arrives #then interrupted remains terminal [0.06ms]
+
+packages/senpi-task/src/store/security.test.ts:
+(pass) TaskRecordStore task id boundary > #given hostile task ids #when public path entry points are called #then ids are rejected before writing [0.65ms]
+(pass) parseTaskRecord persisted boundary > #given malformed optional pid #when records are listed #then diagnostic is typed and record is skipped [0.57ms]
+(pass) parseTaskRecord persisted boundary > #given malformed optional tool allow #when records are listed #then diagnostic is typed and record is skipped [0.53ms]
+(pass) parseTaskRecord persisted boundary > #given invalid enum sentinel #when diagnostic is reported #then raw rejected value is redacted [0.47ms]
+
+ 45 pass
+ 0 fail
+ 1 snapshots, 434 expect() calls
+Ran 45 tests across 9 files. [1.52s]
