+++ .omo/evidence/senpi-task/task-4-category/repair4-test-split/03-typecheck.exit
@@ -1,0 +1,1 @@
+0
