+++ .omo/evidence/senpi-task/task-4-category/repair4-test-split/04-no-excuse-ts-guard.exit
@@ -1,0 +1,1 @@
+0
