+++ .omo/evidence/senpi-task/task-4-category/repair4-test-split/05-loc-check.exit
@@ -1,0 +1,1 @@
+0
