+++ .omo/evidence/senpi-task/task-4-category/repair4-test-split/05-loc-check.log
@@ -1,0 +1,14 @@
+Pure LOC after repair4 split
+245 packages/senpi-task/src/category/resolver.ts
+191 packages/senpi-task/src/category/resolve-category.test.ts
+156 packages/senpi-task/scripts/manual-category-qa.ts
+118 packages/senpi-task/src/category/resolve-category-boundary.test.ts
+116 packages/senpi-task/src/category/openai-categories.ts
+89 packages/senpi-task/src/category/google-categories.ts
+77 packages/senpi-task/src/category/fallback-chains.ts
+69 packages/senpi-task/src/category/types.ts
+45 packages/senpi-task/src/category/anthropic-categories.ts
+44 packages/senpi-task/src/index.ts
+32 packages/senpi-task/src/category/builtins.ts
+31 packages/senpi-task/src/category/kimi-categories.ts
+16 packages/senpi-task/src/category/index.ts
