+++ .omo/evidence/senpi-task/task-4-category/repair4-test-split/06-diff-check.log
@@ -1,0 +1,1 @@
+PASS git diff --check against Todo 4 base for scoped files
