+++ .omo/evidence/senpi-task/task-4-category/repair4-test-split/07-zero-byte-artifact-check.log
@@ -1,0 +1,1 @@
+PASS no zero-byte artifacts under repair4-test-split at 2026-07-06T10:11:08Z
