+++ .omo/evidence/senpi-task/task-4-category/repair4-test-split/baseline-category-tests.log
@@ -1,0 +1,24 @@
+bun test v1.3.14 (0d9b296a)
+
+packages/senpi-task/src/category/resolve-category.test.ts:
+(pass) resolveCategory > #given a builtin category and omo overlay #when resolved #then user config wins and prompt text is appended [1.70ms]
+(pass) resolveCategory > #given a disabled omo category overlay #when resolved #then a disabled result explains the reason [0.09ms]
+(pass) resolveCategory > #given primary model is unavailable and omo fallback exists #when resolved #then delegate-core fallback reaches the registry model [0.25ms]
+(pass) resolveCategory > #given quick primary is unavailable and hardcoded fallback is available #when resolved #then delegate-core fallback chain reaches Anthropic Haiku [0.11ms]
+(pass) resolveCategory > #given ultrabrain primary is unavailable and hardcoded Google fallback is available #when resolved #then delegate-core fallback chain preserves the high variant [0.10ms]
+(pass) resolveCategory > #given no category or fallback model resolves and a system default is available #when resolved #then delegate-core reaches the system default [0.15ms]
+(pass) resolveCategory > #given selected model is absent from registry #when resolved #then unavailable result names attempted and available models [0.10ms]
+(pass) resolveCategory > #given category params in omo overlay #when resolved #then child spec carries generation params and prompt append [0.09ms]
+(pass) resolveCategory > #given a registry model with legal headers #when resolved #then the header-bearing model is accepted [0.08ms]
+(pass) resolveCategory > #given a malformed registry entry #when resolved #then category resolution returns sanitized model_unavailable instead of throwing [0.07ms]
+(pass) resolveCategory > #given malformed truthy find results #when resolved #then category resolution returns sanitized model_unavailable [0.16ms]
+(pass) resolveCategory > #given inherited model identity fields #when resolved #then category resolution rejects them without leaking prototype data [0.08ms]
+(pass) resolveCategory > #given non-array registry availability #when resolved #then category resolution returns sanitized model_unavailable instead of throwing [0.08ms]
+(pass) resolveCategory > #given prototype-shaped category names #when resolved #then they return not_found instead of inherited object values [0.07ms]
+(pass) resolveCategory > #given a custom category description #when resolved #then the resolved result preserves it [0.06ms]
+(pass) builtin category defaults > #given ported builtin defaults #when snapshotted #then all eight category defaults stay pinned [0.39ms]
+
+ 16 pass
+ 0 fail
+ 1 snapshots, 84 expect() calls
+Ran 16 tests across 1 file. [83.00ms]
