+++ packages/senpi-task/src/category/resolve-category.test.ts
@@ -204,132 +204,6 @@ describe("resolveCategory", () => {
     expect(resolved.spec.prompt_append).toEndWith("\n\nEXTRA QUICK CONTEXT")
   })
 
-  test("#given a registry model with legal headers #when resolved #then the header-bearing model is accepted", () => {
-    // given
-    const headerModel = {
-      provider: "openai",
-      id: "gpt-5.4-mini",
-      headers: { "User-Agent": "test" },
-    }
-
-    // when
-    const result = resolveCategory("quick", {}, registry([headerModel]))
-
-    // then
-    const resolved = expectResolved(result)
-    expect(resolved.spec.provider).toBe("openai")
-    expect(resolved.spec.modelId).toBe("gpt-5.4-mini")
-    expect(resolved.spec.model).toBe(headerModel)
-  })
-
-  test("#given a malformed registry entry #when resolved #then category resolution returns sanitized model_unavailable instead of throwing", () => {
-    // given
-    const malformedRegistry = {
-      getAvailable: () => [null],
-      find: () => undefined,
-    }
-
-    // when
-    const result = resolveCategory("quick", {}, malformedRegistry)
-
-    // then
-    expect(result.kind).toBe("model_unavailable")
-    if (result.kind !== "model_unavailable") throw new Error("Expected unavailable result")
-    expect(result.category).toBe("quick")
-    expect(result.attemptedModel).toBe("openai/gpt-5.4-mini")
-    expect(result.availableModels).toEqual([])
-  })
-
-  test("#given malformed truthy find results #when resolved #then category resolution returns sanitized model_unavailable", () => {
-    // given
-    const malformedFindResults = [
-      {},
-      { provider: { secret: "hidden" }, id: ["gpt-5.4-mini"] },
-      { provider: "openai", id: "gpt-5.4-mini", password: "hidden" },
-      { provider: "openai", id: "gpt-5.4-mini", accessToken: "hidden" },
-      { provider: "openai", id: "gpt-5.4-mini", privateToken: "hidden" },
-    ]
-    const availableModel = model("openai", "gpt-5.4-mini")
-
-    // when
-    const results = malformedFindResults.map((findResult) => resolveCategory("quick", {}, {
-      getAvailable: () => [availableModel],
-      find: () => findResult,
-    }))
-
-    // then
-    for (const result of results) {
-      expect(result.kind).toBe("model_unavailable")
-      if (result.kind !== "model_unavailable") throw new Error(`Expected unavailable result, got ${result.kind}`)
-      expect(result.attemptedModel).toBe("openai/gpt-5.4-mini")
-      expect(result.availableModels).toEqual(["openai/gpt-5.4-mini"])
-      expect(JSON.stringify(result)).not.toContain("hidden")
-    }
-  })
-
-  test("#given inherited model identity fields #when resolved #then category resolution rejects them without leaking prototype data", () => {
-    // given
-    const availableModel = model("openai", "gpt-5.4-mini")
-    const inheritedIdentityModel: object = Object.create({
-      provider: "openai",
-      id: "gpt-5.4-mini",
-      privateToken: "hidden",
-    })
-
-    // when
-    const result = resolveCategory("quick", {}, {
-      getAvailable: () => [availableModel],
-      find: () => inheritedIdentityModel,
-    })
-
-    // then
-    expect(result.kind).toBe("model_unavailable")
-    if (result.kind !== "model_unavailable") throw new Error(`Expected unavailable result, got ${result.kind}`)
-    expect(result.attemptedModel).toBe("openai/gpt-5.4-mini")
-    expect(result.availableModels).toEqual(["openai/gpt-5.4-mini"])
-    expect(JSON.stringify(result)).not.toContain("hidden")
-  })
-
-  test("#given non-array registry availability #when resolved #then category resolution returns sanitized model_unavailable instead of throwing", () => {
-    // given
-    const malformedAvailableResults = [
-      null,
-      { 0: model("openai", "gpt-5.4-mini"), length: 1 },
-      "openai/gpt-5.4-mini",
-    ]
-
-    // when
-    const results = malformedAvailableResults.map((availableResult) => resolveCategory("quick", {}, {
-      getAvailable: () => availableResult,
-      find: () => model("openai", "gpt-5.4-mini"),
-    }))
-
-    // then
-    for (const result of results) {
-      expect(result.kind).toBe("model_unavailable")
-      if (result.kind !== "model_unavailable") throw new Error(`Expected unavailable result, got ${result.kind}`)
-      expect(result.attemptedModel).toBe("openai/gpt-5.4-mini")
-      expect(result.availableModels).toEqual([])
-    }
-  })
-
-  test("#given prototype-shaped category names #when resolved #then they return not_found instead of inherited object values", () => {
-    // given
-    const models = registry([model("openai", "gpt-5.4-mini")])
-
-    // when
-    const results = ["__proto__", "toString", "hasOwnProperty"].map((category) =>
-      resolveCategory(category, {}, models)
-    )
-
-    // then
-    expect(results.map((result) => result.kind)).toEqual(["not_found", "not_found", "not_found"])
-    for (const result of results) {
-      if (result.kind !== "not_found") throw new Error(`Expected not_found result, got ${result.kind}`)
-      expect(result.availableCategories).toContain("quick")
-    }
-  })
-
   test("#given a custom category description #when resolved #then the resolved result preserves it", () => {
     // given
     const models = registry([model("openai", "custom-model")])
